modified threshold and nj param calculation that outputs for n training days

+++ TrajectoryClustering_Traclus.py
@@ -55,9 +55,7 @@ def TrajectoryClustering_Traclus(traFileCreation, day_to_analyze, year_to_analyz
     currentDay = dayArray[0]
     pointerMonth = 0
     pointerDay = 0
-    daysToChoose = 365 #30 for original experiment
-    cycle_loop = 8
-    cycle_counter = 0
+    daysToChoose = 40 #30 for original experiment
     num_days = 4 # this parameter sets total days to be considered
     n_parameter = 3 # this parameter sets how many previous training days
     annotate_testdays = np.empty([0]) 
@@ -104,8 +102,9 @@ def TrajectoryClustering_Traclus(traFileCreation, day_to_analyze, year_to_analyz
         counter += 1
     
         currentYear, currentMonth, currentDay, pointerMonth, pointerDay = calendarFunction(currentYear, currentMonth, currentDay, pointerMonth, pointerDay)    
+    
     #####################################################################################################
-#    print(list_day)
+    print(list_day)
 #    print(len(list_day))
 #    print("start tra")
 #    os.chdir(root_dir)
@@ -132,7 +131,7 @@ def TrajectoryClustering_Traclus(traFileCreation, day_to_analyze, year_to_analyz
             annotate_testdays = np.append(annotate_testdays, list_day[partition_num])
             if len(frames) > 0:
                 concatDay = frames.sort_values(by='TrackID', ascending=True)
-                allFrames = allFrames.append(concatDay)
+#                allFrames = allFrames.append(concatDay,  ignore_index= True)
                 concatDay = concatDay.reset_index(drop=True)
             filename = "test_day_" + str(counter_list_day) + "_" + str(day_to_analyze) + ".tra"
             os.chdir(root_dir)            
@@ -147,7 +146,7 @@ def TrajectoryClustering_Traclus(traFileCreation, day_to_analyze, year_to_analyz
             days_in_partition = 0            
         counter_list_day += 1
 #    print(date_verifier) # verifying the dates trained sequentially
-
+    
     ####################################Visualizing Multiple##########################################
 #    print(annotate_testdays)
     visualize_num = 5
@@ -159,14 +158,14 @@ def TrajectoryClustering_Traclus(traFileCreation, day_to_analyze, year_to_analyz
     nj_parameter_array = np.empty([0])
     threshold_array = np.empty([0])
     
-    ## nj param
+    ##################################### nj param ##############################
     os.chdir(root_dir)
     counter_nj = 0
     while counter_nj < ( len(list_day) - n_parameter ):
         
         filename = "test_day_" + str(counter_nj) + "_" + str(day_to_analyze) + '.txt'
         image_filename = "test_day_" + str(counter_nj) + "_" + str(day_to_analyze) + '.jpg'
-         
+        
         # 3-> Thursday
         major_track, end, lines_ori = ReadTraclusExport(filename)
         temp_major_track = major_track
@@ -181,65 +180,77 @@ def TrajectoryClustering_Traclus(traFileCreation, day_to_analyze, year_to_analyz
         cluster_line = LinesConstruct(temp_major_track)
         Visualizer(cluster_line, colors_for_lines, image_filename, root_dir, img_dir)     
         
-#        print(temp_major_track)
         # allFrames 
-        tracks = TrajectoryID_Extraction(major_track, allFrames, end, lines_ori)
-        # tracks -> all tracks under the cluster
-        tracksLine = LinesConstruct(tracks)
+#        tracks = TrajectoryID_Extraction(major_track, allFrames, end, lines_ori)
+#        # tracks -> all tracks under the cluster
+#        tracksLine = LinesConstruct(tracks)
         #tracksline -> extract X & Y(a python list)
-    #    print(filename)
-        nj_param = nj_training_parameter(tracks, temp_major_track)
-        nj_parameter_array = np.append(nj_parameter_array, nj_param)
+        
+#        nj_param = nj_training_parameter(tracks, temp_major_track)
+#        nj_parameter_array = np.append(nj_parameter_array, nj_param)
         counter_nj += 1
     nj_param = np.mean(nj_parameter_array)
-    probability_array = np.empty([0])
+    #############################################################################    
+
     
-    
-    # array for second level clustering
-    array_second_representative = np.empty([0])
-    # threshold calculation
+    ########################## threshold calculation ###########################
+    partition_start = 0
+        
     while counter_new_day < ( len(list_day) - n_parameter ):
-#        filename = "Day_6_" + str(counter_new_day) + ".txt"
+        # Cluster/representative track
+        os.chdir(root_dir)
         filename = "test_day_" + str(counter_new_day) + "_" + str(day_to_analyze) + '.txt'
-#        filename = "All_days_" + str(day_to_analyze) + ".txt"        
-        # 3-> Thursday
         major_track, end, lines_ori = ReadTraclusExport(filename)
-        array_second_representative = np.append(array_second_representative, major_track)
-        temp_major_track = major_track
-        major_track = major_track[['X', 'Y']]
-        major_track = major_track.as_matrix()
+        
+        # Training track
+        partition_num = partition_start
+        counter = 0
+        os.chdir(traj_dir)
+        while counter < n_parameter: 
+            filename = list_day[partition_num] + ".txt"
+            allFrames = allFrames.append(pd.read_table(filename, delimiter=' ', header=None, names=col_names), ignore_index= True)
+            print(filename)     
+            partition_num += 1
+            counter += 1
+        allFrames = allFrames[['TrackID', 'X', 'Y']]
         temp_threshold_array = np.empty([0])
-#        print(temp_major_track)
-        # allFrames 
         tracks = TrajectoryID_Extraction(major_track, allFrames, end, lines_ori)
-        # tracks -> all tracks under the cluster
-        tracksLine = LinesConstruct(tracks)
+        # tracks -> all tracks under the cluster        
         #tracksline -> extract X & Y(a python list)
-#        tracks = tracks[['X', 'Y']]
-#        tracks = tracks.as_matrix()
-#        print("similarity")
-        threshold_results = calculateSimilarity(temp_major_track, tracksLine, nj_param)
-        temp_threshold_array = np.append(temp_threshold_array, threshold_results)
-          
+        
+        nj_param = nj_training_parameter(allFrames, major_track)
+        nj_parameter_array = np.append(nj_parameter_array, nj_param)
+
+        
+        threshold_results = calculateSimilarity(temp_major_track, allFrames, nj_param)
+        temp_threshold_array = np.append(temp_threshold_array, threshold_results)         
         threshold_array = np.append(threshold_array, min(temp_threshold_array))
 
+        partition_start += 1
         counter_new_day += 1
-        
+     ###########################################################################
+       
         
 
-#    print(threshold_array) 
-    threshold_array = np.sort(threshold_array)
-    threshold_array = np.percentile(threshold_array, 50)
+    ############################## Output File Section #############################
+    os.chdir(root_dir)
     file = "threshold_array_" + str(day_to_analyze) + ".txt" 
-    output = open(file, 'w')
-    output.write(str(threshold_array))
+    if os.path.isfile(file) == False:
+        output = open(file, 'w')
+    else:
+        output = open(file, 'a')
+    for days in threshold_array:  
+        output.write(str(days) + "\n")
     output.close()
     
     file = "nj_param_" + str(day_to_analyze) + ".txt"
-    output = open(file, 'w') 
-    output.write(str(nj_param))
+    if os.path.isfile(file) == False:
+        output = open(file, 'w')
+    else:
+        output = open(file, 'a')
+    for days in nj_parameter_array:     
+        output.write(str(days) + "\n")
     output.close()
-    
     
     file = "test_days_list_" + str(day_to_analyze) +".txt"
     if os.path.isfile(file) == False:
@@ -250,10 +261,11 @@ def TrajectoryClustering_Traclus(traFileCreation, day_to_analyze, year_to_analyz
     for days in annotate_testdays:
         print(str(days))        
         output.write(str(days) + "\n")
-#    print("threshold array")
-#    print(threshold_array)
-    ############################################################################################################
-#    print(allFrames)
+    #################################################################################
+        
+        
+        
+        
 #userParam = sys.argv
 #export_file = bool(userParam[1])
 #day_in_week = int(userParam[2])
@@ -281,14 +293,20 @@ def TrajectoryClustering_Traclus(traFileCreation, day_to_analyze, year_to_analyz
 #        anomaly_detection(export_file, day_in_week, num_days, year, root_dir, data_dir, image_dir)
 #except:
 #    print("Invalid Input. Make sure it's TrajectoryClustering_Traclus.py [bool(True|False)] [int(0~6)]")
-#TrajectoryClustering_Traclus(True, 0, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
+TrajectoryClustering_Traclus(True, 0, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
 #TrajectoryClustering_Traclus(True, 1, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
 #TrajectoryClustering_Traclus(True, 2, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
 #TrajectoryClustering_Traclus(True, 3, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
 #TrajectoryClustering_Traclus(True, 4, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
 #TrajectoryClustering_Traclus(True, 5, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
 #TrajectoryClustering_Traclus(True, 6, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
-anomaly_detection(True, 0, 60, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
+#anomaly_detection(True, 0, 60, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
+#anomaly_detection(True, 1, 365, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
+#anomaly_detection(True, 2, 365, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
+#anomaly_detection(True, 3, 365, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
+#anomaly_detection(True, 4, 365, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
+#anomaly_detection(True, 5, 365, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
+#anomaly_detection(True, 6, 365, 2012, 'E:\Documents\MMU Studies\Python Scripts', 'E:\Documents\MMU Studies\Python Scripts\Track LOST dataset', "E:\Documents\MMU Studies\Python Scripts\Trajectories")
 
 
 #TrajectoryClustering_Traclus(True, 0, 2012)
